backend/Django/api_cidades/requirements.txt
django
djangorestframework
dj-database-url
psycopg2-binary
django-cors-headers
drf-yasg
djangorestframework-simplejwt[crypto]
gunicorn
